+++ packages/insomnia/package.json
@@ -17,15 +17,15 @@
   "author": "Kong <office@konghq.com>",
   "main": "src/main.min.js",
   "scripts": {
-    "watch:hidden-window": "esbuild --watch --bundle --platform=browser --sourcemap --format=cjs --target=esnext --outfile=./src/renderers/hidden-browser-window/build/index.js ./src/renderers/hidden-browser-window/index.ts",
     "build": "npm run build:sr && npm run build:app",
     "build:app": "esr --cache ./scripts/build.ts --noErrorTruncation",
     "build:main.min.js": "cross-env NODE_ENV=development esr esbuild.main.ts",
     "build:sr": "esr esbuild.sr.ts && tsc --build tsconfig.build.sr.json",
     "lint": "eslint . --ext .js,.ts,.tsx --cache",
     "package": "npm run build:app && cross-env USE_HARD_LINKS=false electron-builder build --config electron-builder.config.js",
-    "start": "concurrently -n dev,app --kill-others \"npm run start:dev-server\" \"npm run start:electron\"",
+    "start": "concurrently -n browser,hidden-browser,main --kill-others \"npm run start:dev-server\"  \"npm run start:hidden-window\" \"npm run start:electron\"",
     "start:dev-server": "vite dev",
+    "start:hidden-window": "esbuild --watch --bundle --platform=browser --sourcemap --format=cjs --target=esnext --outfile=./src/renderers/hidden-browser-window/build/index.js ./src/renderers/hidden-browser-window/index.ts",
     "start:electron": "cross-env NODE_ENV=development esr esbuild.main.ts && electron --inspect=5858 .",
     "test": "jest",
     "electron:dev-build": "electron ./build/main.min.js",
